Move JSON specific test from synced collection tests to to JSON tests.
