+++ tests/test_synced_collections/synced_collection_test.py
@@ -6,7 +6,6 @@ from copy import deepcopy
 
 import pytest
 
-from signac.core.synced_collections.collection_json import JSONDict
 from signac.core.synced_collections.synced_collection import SyncedCollection
 from signac.errors import InvalidKeyError, KeyTypeError
 
@@ -387,17 +386,6 @@ class SyncedDictTest(SyncedCollectionTest):
             synced_collection[key] = testdata
             assert key in synced_collection
             assert synced_collection[key] == testdata
-
-    # The following test tests the support for non-str keys
-    # for JSON backend which will be removed in version 2.0.
-    # See issue: https://github.com/glotzerlab/signac/issues/316.
-    def test_keys_non_str_valid_type(self, synced_collection, testdata):
-        if isinstance(synced_collection, JSONDict):
-            for key in (0, None, True):
-                with pytest.deprecated_call(match="Use of.+as key is deprecated"):
-                    synced_collection[key] = testdata
-                assert str(key) in synced_collection
-                assert synced_collection[str(key)] == testdata
 
     def test_keys_invalid_type(self, synced_collection, testdata):
         class A:
