+++ tests/test_synced_collections/test_json_collection.py
@@ -53,6 +53,16 @@ class TestJSONDict(JSONCollectionTest, SyncedDictTest):
 
     _collection_type = JSONDict
 
+    # The following test tests the support for non-str keys
+    # for JSON backend which will be removed in version 2.0.
+    # See issue: https://github.com/glotzerlab/signac/issues/316.
+    def test_keys_non_str_valid_type(self, synced_collection, testdata):
+        for key in (0, None, True):
+            with pytest.deprecated_call(match="Use of.+as key is deprecated"):
+                synced_collection[key] = testdata
+            assert str(key) in synced_collection
+            assert synced_collection[str(key)] == testdata
+
 
 class TestJSONList(JSONCollectionTest, SyncedListTest):
